% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_.R
\name{calc_max_age}
\alias{calc_max_age}
\title{Calculate maximum age from natural mortality}
\usage{
calc_max_age(natural_mortality)
}
\arguments{
\item{natural_mortality}{A vector of positive real numbers specifying the
natural mortality for a given group.}
}
\value{
A vector of estimates of maximum age, where the length and names of the
returned vector are determined by the input vector.
}
\description{
Using Hamel and Cope (2022), the maximum observed age can serve as a proxy
for longevity of marine fishes where natural mortality equals 5.4 divided by
the maximum age. This function inverts that formula to determine the maximum
age that you should see in your population given the estimate of natural
mortality.
}
