% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_rfishbase_version}
\alias{check_rfishbase_version}
\title{Check if the Installed Version of \code{rfishbase} is Up to Date}
\usage{
check_rfishbase_version()
}
\description{
This function compares the installed version of the \code{rfishbase} package
with the latest version available on CRAN. It prints a message if the
installed version is outdated.
}
\examples{
check_rfishbase_version()

}
