% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishprior-package.R
\docType{package}
\name{fishprior-package}
\alias{fishprior}
\alias{fishprior-package}
\title{fishprior: Get Priors for Life-History Traits of Fishes}
\description{
To learn more about fishprior, start with the vignettes found in
\code{browseVignettes(package = "fishprior")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/noaa-fims/fishprior}
  \item Report bugs at \url{https://github.com/noaa-fims/fishprior/issues}
}

}
\author{
\strong{Maintainer}: Kelli F. Johnson \email{kelli.johnson@noaa.gov} (\href{https://orcid.org/0000-0002-5149-451X}{ORCID})

Authors:
\itemize{
  \item Jennifer S. Bigman, (\href{https://orcid.org/0000-0001-8070-3061}{ORCID})
}

}
\keyword{internal}
