% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fishlife_traits.R
\name{get_fishlife_traits}
\alias{get_fishlife_traits}
\title{Get traits for a species from the FishLife package}
\usage{
get_fishlife_traits(
  species,
  keep_regexp = "age|fecundity|length_|mortality|weight|growth"
)
}
\arguments{
\item{species}{A string specifying the scientific name for the species of
interest. This function will only accept a single string. See the examples
for how to get data on multiple species at once.}

\item{keep_regexp}{A string specifying a regular expression to search for.
The default searches for multiple words using \code{"|"} as a separator. This
default will return nine traits related to life history.}
}
\value{
A data frame with four columns is returned. The first column, \code{name}, stores
the scientific/latin name for the species. The same value that was used in
the \code{species} input argument. The second column, \code{trait}, stores the name of
the life-history trait. Most of these are start with log( to specify that the
values are in natural log space. The third and fourth columns, \code{mean} and
\code{se}, store the mean and the standard error of the mean for the given trait.
}
\description{
Get life-history traits for a given species from the FishLife package. Look
at the example for how to get data on multiple species at once. This function
depends on the \code{FishBase_and_Morphometrics} data object from the FishLife
package.
}
\examples{
species_list <- c(
  "Merluccius merluccius",
  "Gadus chalcogrammus",
  "Anoplopoma fimbria"
)
get_fishlife_traits(species_list[1])
# Get just natural mortality information
get_fishlife_traits(species_list[1], keep_regexp = "mort")
# Expect error because the species PacFIN does not exist
\dontrun{
get_fishlife_traits("PacFIN")
}
# Expect error because the function is not vectorized
\dontrun{
get_fishlife_traits(species_list[1:2])
}
# Example of how to get vectorized output and save it
traits <- purrr::map_df(species_list, get_fishlife_traits)
\dontrun{
utils::write.csv(traits, file = "example_traits_FishLife.csv")
}
}
