% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_growth_.R
\name{make_growth_coefficient_prior}
\alias{make_growth_coefficient_prior}
\title{Make a prior for the growth coefficient}
\usage{
make_growth_coefficient_prior(data, type = c("informative", "diffuse"))
}
\arguments{
\item{data}{A data frame for a single species because you might
need more than one trait to create a prior.}

\item{type}{A character string indicating the type of prior. The default
is "informative".}
}
\description{
TODO: Provide more details on the parameter.
}
