% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mortality_prior.R
\name{make_informative_mortality_prior}
\alias{make_informative_mortality_prior}
\title{Make an information prior for natural mortality}
\usage{
make_informative_mortality_prior(log_maximum_age)
}
\arguments{
\item{log_maximum_age}{A numeric value representing the maximum age of the
fish species on the log scale. This value must be greater than 0.}
}
\value{
An informative lognormal distribution is returned using the \code{prior} class.
}
\description{
The Hamel--Cope prior for natural mortality is thought to be a good,
informative prior for natural mortality. It is based on the maximum age of
a fish species.
}
\details{
The Hamel--Cope prior is based on the assumption that natural mortality is
lognormally distributed. The mean of the lognormal distribution is
calculated as 5.40 divided by the maximum age of the fish species. The
standard deviation of the lognormal distribution is set to 0.31.
}
\examples{
prior_35 <- make_informative_mortality_prior(log(35))
exp(get_parameters(prior_35)[["mean_log"]])
traits_example |>
  dplyr::filter(
    trait == "log(age_max)"
  ) |>
  dplyr::mutate(
    natural_mortality = purrr::map(
      mean,
      make_informative_mortality_prior
    )
  )
}
\keyword{informative}
\keyword{mortality}
\keyword{prior}
