% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mortality_.R
\name{make_mortality_prior}
\alias{make_mortality_prior}
\title{Make a prior for natural mortality}
\usage{
make_mortality_prior(data, type = c("diffuse", "informative"))
}
\arguments{
\item{data}{A data frame returned by \code{\link[=get_fishlife_traits]{get_fishlife_traits()}} containing
life-history information for at least one fish species.}

\item{type}{A character string indicating the type of prior to be created.
Either \code{"diffuse"} or \code{"informative"}. The default is \code{"diffuse"}.}
}
\value{
A tibble is returned that contains three columns, \code{"name"} from the input
data frame with the species name, \code{"parameter"} specifying which parameter
the prior is for, and \code{"prior"} that contains the S4 class \code{prior}.
}
\description{
Make a prior for natural mortality using either an informative or
diffuse prior. The informative prior is based on the Hamel--Cope prior,
which is based on the maximum age of a fish species. The diffuse prior is
based on the estimated natural mortality of a fish species.
}
\keyword{mortality}
\keyword{prior}
