% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase.R
\name{map_trait_fb}
\alias{map_trait_fb}
\title{Summarize FishBase Trait Data Tables, Traits, and Names}
\usage{
map_trait_fb()
}
\description{
This function takes no input, but is just useful for identifying the source
of trait info from FishBase and the corresponding table names.
}
