% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mortality_decay.R
\name{plot_mortality_decay}
\alias{plot_mortality_decay}
\title{Plot the decay of natural mortality by age}
\usage{
plot_mortality_decay(x)
}
\arguments{
\item{x}{A vector of positive real numbers specifying natural mortality
for each desired group. The vector should be a named vector if it has a
length of great than one, e.g., \code{c("female" = 0.33, "male" = 0.2)}.}
}
\value{
A {ggplot2} object is returned.
}
\description{
Starting at age zero to the maximum theoretical age (based on natural
mortality), the decay of mortality (y axis) is shown by age (years).
Lines are colored by group, where groups are defined by the names of
the input vector.
}
\examples{
plot_mortality_decay(0.2)
plot_mortality_decay(seq(0.2, 0.4, by = 0.1))
plot_mortality_decay(c("female" = .2, "male" = .4))
}
\seealso{
\itemize{
\item \code{calc_max_age()} for how the maximum age is calculated
}
}
