% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_fishlife_traits.R
\name{summarize_fishlife_traits}
\alias{summarize_fishlife_traits}
\title{Summarize FishLife trait data into log-transformed traits}
\usage{
summarize_fishlife_traits(
  species,
  keep_regexp = "age|fecundity|length_|mortality|weight|growth"
)
}
\arguments{
\item{species}{A vector of strings specifying the scientific names for the
species of interest.}

\item{keep_regexp}{A string specifying a regular expression to search for.
The default searches for multiple words using \code{"|"} as a separator. This
default will return nine traits related to life history.}
}
\value{
A data frame with four columns is returned. The first column, \code{name}, stores
the scientific/latin name for the species. The same value that was used in
the \code{species} input argument. The second column, \code{trait}, stores the name of
the life-history trait. Most of these are start with log( to specify that the
values are in natural log space. The third and fourth columns, \code{mean} and
\code{se}, store the mean and the standard error of the mean for the given trait.
}
\description{
FishLife already summarizes these traits for us so this function is just
returning a table for a given species from the FishLife package. This
function depends on the \code{FishBase_and_Morphometrics} data object from the
FishLife package.
}
\examples{
species_list <- c(
  "Merluccius merluccius",
  "Gadus chalcogrammus",
  "Anoplopoma fimbria"
)
summarize_fishlife_traits(species_list)
# Get just natural mortality information
summarize_fishlife_traits(species_list, keep_regexp = "mort")
# Expect error because the species PacFIN does not exist
\dontrun{
summarize_fishlife_traits("PacFIN")
}
}
\seealso{
\itemize{
\item \code{\link[=summarize_fishbase_traits]{summarize_fishbase_traits()}}
}
}
