% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits_example.R
\docType{data}
\name{traits_example}
\alias{traits_example}
\title{Example trait information for three species}
\format{
\subsection{\code{who}}{

A data frame with 27 rows and 4 columns:
\describe{
\item{name}{The latin/scientific name of the species}
\item{trait}{A string in an R-code style indicating which trait the values
are for}
\item{mean, se}{The mean and standard error of the given trait}
}
}
}
\source{
\url{https://james-thorson-noaa.github.io/FishLife/}
}
\usage{
traits_example
}
\description{
An example of trait information obtained from the FishLife package for three
species, European Hake, Alaska pollock, and sablefish.
}
\keyword{datasets}
