% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_trait_names.r
\name{translate_trait_names}
\alias{translate_trait_names}
\title{Change FishBase trait name to snake-case descriptions}
\usage{
translate_trait_names(trait)
}
\arguments{
\item{trait}{A character vector of trait names from FishBase.}
}
\value{
A character vector of trait names in snake_case format rather than
the format found in FishBase, which is often shorted versions of the trait
names.
}
\description{
The resultant trait names are the snake-case names that are used in
FishLife and are preceded by \verb{log(} to indicate that the traits are in
log space.
}
\examples{
translate_trait_names("Loo")
translate_trait_names("K")
translate_trait_names(c("Loo", "K", "M"))
}
