% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase.R
\name{summarize_fishbase_traits}
\alias{summarize_fishbase_traits}
\title{Summarize FishBase trait data into log-transformed traits}
\usage{
summarize_fishbase_traits(fb)
}
\arguments{
\item{fb}{A tibble returned by \code{\link[=get_fishbase_traits]{get_fishbase_traits()}}.}
}
\value{
A tibble grouped by Species and trait with log-transformed stats.
}
\description{
Summarize FishBase trait data into log-transformed traits
}
